# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Legacy Sports API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schools:
    get:
      operationId: listSchools
      tags: [schools]
      summary: List all schools
      responses:
        "200":
          description: List of schools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/School"
    post:
      operationId: createSchool
      tags: [schools]
      summary: Create a school
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolInput"
      responses:
        "201":
          description: School created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
  /schools/{id}:
    get:
      operationId: getSchool
      tags: [schools]
      summary: Get school by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: School
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
    put:
      operationId: updateSchool
      tags: [schools]
      summary: Update school
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolInput"
      responses:
        "200":
          description: Updated school
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
    delete:
      operationId: deleteSchool
      tags: [schools]
      summary: Delete school
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /schools/{id}/pause:
    post:
      operationId: pauseSchool
      tags: [schools]
      summary: Pause or unpause a school
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolPauseInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
...
  /sports:
    get:
      operationId: listSports
      tags: [sports]
      summary: List all sport configs
      responses:
        "200":
          description: Sports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SportConfig"
    post:
      operationId: upsertSport
      tags: [sports]
      summary: Create or update sport config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SportInput"
      responses:
        "200":
          description: Sport config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SportConfig"
  /sports/{id}:
    delete:
      operationId: deleteSport
      tags: [sports]
      summary: Delete sport config
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    School:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        phone:
          type: string
        whatsappNumber:
          type: string
        address:
          type: string
        email:
          type: string
        ownerName:
          type: string
        ownerPhone:
          type: string
        ownerWhatsapp:
          type: string
        ownerEmail:
          type: string
        principalName:
          type: string
        logoUrl:
          type: string
        primaryColor:
          type: string
        isPaused:
          type: boolean
        pauseMessage:
          type: string
        pausedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, code]
    SchoolInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        code:
          type: string
        phone:
          type: string
        whatsappNumber:
          type: string
        address:
          type: string
        email:
          type: string
        ownerName:
          type: string
        ownerPhone:
          type: string
        ownerWhatsapp:
          type: string
        ownerEmail:
          type: string
        principalName:
          type: string
        primaryColor:
          type: string
        logoUrl:
          type: string
    SchoolPauseInput:
      type: object
      required: [paused]
      properties:
        paused:
          type: boolean
        message:
          type: string
...
    SportField:
      type: object
      required: [key, label, type]
      properties:
        key:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [int, float, text]
        section:
          type: string
        min:
          type: number
        auto:
          type: boolean
        placeholder:
          type: string
    SportConfig:
      type: object
      properties:
        id:
          type: integer
        sportName:
          type: string
        icon:
          type: string
        fieldsJson:
          type: array
          items:
            $ref: "#/components/schemas/SportField"
        updatedAt:
          type: string
      required: [id, sportName]
    SportInput:
      type: object
      required: [sportName]
      properties:
        sportName:
          type: string
        icon:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/SportField"
...